Break up manifest.go
